ATF-9: Moved FeedWriterBase and DefaultFeedWriter classes to impl package
